feat: add bspnetConfigCases to challenge cycle test
